refs #79 repair step => job1 (launch one job2 for each user) => job2 (scan for one user)
